Replace GitHistoryUtils.history with collectTimedCommits
history method is really slow when the repo contains a lots of commits
if the method parameters does not specify the number of commit history
this plugin wants. Also, most of places that use history method does
not require all commit info. Therefore, I shifted to use
collectTimedCommits. And then, if it requires other commit info, use
history with the parameter that limit the number of commit.